requirements.txt
astroid==2.2.0
autopep8==1.4.3
Django==2.1.7
djangorestframework==3.9.2
isort==4.3.9
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
Pillow==5.4.1
pkg-resources==0.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pylint==2.3.0
pytz==2018.9
six==1.12.0
typed-ast==1.2.0
wrapt==1.11.1
social-auth-app-django==3.1.0
social-auth-core==3.1.0
